migrate `misc.applyMixins` into `@base/utils/internal`
