+++ cocos/asset/assets/asset.jsb.ts
@@ -24,7 +24,8 @@
 
 import { cclegacy } from '@base/global';
 import { js } from '@base/utils';
-import { _decorator, path, jsbUtils, CallbacksInvoker, applyMixins } from '../../core';
+import { applyMixins } from '@base/utils/internal';
+import { _decorator, path, jsbUtils, CallbacksInvoker } from '../../core';
 import { getUrlWithUuid } from '../asset-manager/helper';
 import { patch_cc_Asset } from '../../native-binding/decorators';
 import type { Asset as JsbAsset } from './asset';
