+++ cocos/base/utils/src/internal-index.ts
@@ -8,3 +8,13 @@ export function isPlainEmptyObj (obj): boolean {
     }
     return js.isEmptyObject(obj);
 }
+
+export function applyMixins (derivedCtor: any, baseCtors: any[]): void {
+    baseCtors.forEach((baseCtor) => {
+        Object.getOwnPropertyNames(baseCtor.prototype).forEach((name) => {
+            if (name !== 'constructor') {
+                Object.defineProperty(derivedCtor.prototype, name, Object.getOwnPropertyDescriptor(baseCtor.prototype, name)!);
+            }
+        });
+    });
+}
