+++ cocos/core/utils/misc.ts
@@ -268,33 +268,6 @@ export function radiansToDegrees (angle): number {
     return angle * macro.DEG;
 }
 
-/**
- * @engineInternal
- */
-export function applyMixins (derivedCtor: any, baseCtors: any[]): void {
-    baseCtors.forEach((baseCtor) => {
-        Object.getOwnPropertyNames(baseCtor.prototype).forEach((name) => {
-            if (name !== 'constructor') {
-                Object.defineProperty(derivedCtor.prototype, name, Object.getOwnPropertyDescriptor(baseCtor.prototype, name)!);
-            }
-        });
-    });
-}
-
-// if (TEST) {
-//     // editor mocks using in unit tests
-//     if (typeof Editor === 'undefined') {
-//         window.Editor = {
-//             UuidUtils: {
-//                 NonUuidMark: '.',
-//                 uuid () {
-//                     return '' + ((new Date()).getTime() + Math.random());
-//                 },
-//             },
-//         };
-//     }
-// }
-
 cclegacy.misc = {
     BUILTIN_CLASSID_RE,
     BASE64_VALUES,
